requirements.txt
torch
torchvision
pytorch-metric-learning
faiss-cpu
scipy
numpy
pillow
tqdm
streamlit
